Move files and adjust spaces to 2
