+++ assets/js/sidebar.js
@@ -10,11 +10,11 @@
  */
 
 jQuery(document).ready(function(){
-    $('.sidebar-toggle').click(function(){
-        if (~document.cookie.indexOf('sonata_sidebar_hide=1')) {
-            return document.cookie = 'sonata_sidebar_hide=0;path=/';
-        }
+  $('.sidebar-toggle').click(function(){
+    if (~document.cookie.indexOf('sonata_sidebar_hide=1')) {
+      return document.cookie = 'sonata_sidebar_hide=0;path=/';
+    }
 
-        document.cookie = 'sonata_sidebar_hide=1;path=/';
-    });
+    document.cookie = 'sonata_sidebar_hide=1;path=/';
+  });
 });
